:truck: Rename register components to more representative names
