+++ src/modules/Auth/components/RegisterForm/RecoveryPhraseSafeKeepingScreen/RecoveryPhraseSafeKeepingScreen.styles.js
