+++ src/modules/Auth/components/RegisterForm/RecoveryPhraseStrengthSelect/RecoveryPhraseStrengthSelect.js
