+++ src/modules/Auth/components/RegisterForm/RecoveryPhraseStrengthSelect/RecoveryPhraseStrengthSelect.styles.js
